\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma de dos números}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
  \item{x}{ Primer sumando }
  \item{y}{ Segundo sumando }
}
\description{
Función que suma 2 números recibidos como argumento
}
\author{
Diego Castro
}
\examples{
suma.dos.numeros(1, 1)
}
