\name{tablasOperacion}
\alias{tablasOperacion}
\title{Recupera las tablas de una operaci贸n del api del INE}
\usage{
tablasOperacion(operacionId)
}
\arguments{
  \item{operacionId}{ Identificador de la operaci贸n de la cual se quieren recuperar sus tablas }
}
\description{
Funci贸n que recupera las tablas de una operaci贸n del api del INE
}
\author{
Diego Castro
}
\examples{
df <- tablasOperacion(22)
}
